test: clock: convert from deterministic values
Using vlc_tick_now() makes the test depend on when it is run. Instead,
provide fixed values to ensure the same inputs in the test.

It can be noted that using VLC_TICK_0 instead of VLC_TICK_0 + 15000ms
makes the test fail though, because conversion in sent to the future.

Co-authored-by: Thomas Guillem <thomas@gllm.fr>