TODO: clock: disable clock start check
The main issue highlighted by this check was that the clock was being
used before the input is actually started, which could happen when a
flush has been called on the decoder, unblocking it before the input is
started.

This assertion should probably be removed since we cannot test the whole
clock startup behaviour with it.