qml: show media library progress bar when media library thread is busy
In addition to media library's own idle scanning status, also show
the progress bar when the media library thread is busy.

This is useful when media library is overloaded, such that it can not
complete the queued tasks in human moment. This can happen, for example,
when a lot of items are added to a playlist.